\name{getPerlScript}
\alias{getPerlScript}
\title{Find a Perl Script}
\description{
Locates the fully qualified name of the perl script
in this package.
This is used to bootstrap the package when it is loaded. 
}
\usage{
getPerlScript(name)
}
\arguments{
  \item{name}{the name of the perl script}
}
\details{

}
\value{
 The fully qualified name of the Perl script.
}
\references{\url{http://www.omegahat.org/RSPerl}}
\author{Duncan Temple Lang}
\seealso{
\code{\link{.PerlFile}}
\code{\link{.PerlInit}}
}
\examples{

}
\concept{Perl}
\concept{Inter-system interface}
\keyword{interface}
