\name{.PerlTypes}
\alias{.PerlTypes}
\non_function{}
\title{Enumerated list of Perl data type constants}
\description{
 This provides a mapping from the names of Perl data types
to their internal representations that can be passed
to the C routines that implement the interface.
This is a named integer vector with the names
giving the human readable and usable form of the Perl
data type and the values giving the corresponding internal
representations.
}
\keyword{programming}
\concept{Inter-system interface}
\concept{Perl}
\concept{marshall}